pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gravity-dca-bot"
version = "0.1.0"
description = "A configurable DCA bot for GRVT perpetual markets."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "grvt-pysdk>=0.2.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
gravity-dca = "gravity_dca.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
